Move LTO symbol export calculation from backends to cg_ssa
